refactor: relocate workflowsview to the shared package (#1610)
