+++ packages/shared/views/WorkflowsView/components/Workflows/types.ts
@@ -1,4 +1,4 @@
-import type { WorkflowEntity } from "@/views/WorkflowsView/types";
+import type { WorkflowEntity } from "@repo/shared/views/WorkflowsView/types";
 import { type Table } from "@tanstack/react-table";
 
 export interface Props {
