+++ pages/data/workflows/index.tsx
@@ -1,8 +1,8 @@
 import { getPageMeta } from "@/common/meta/utils";
 import { config } from "@/config/config";
-import { WorkflowsView } from "@/views/WorkflowsView/workflowsView";
 import { Main as DXMain } from "@databiosphere/findable-ui/lib/components/Layout/components/Main/main.styles";
 import { EntityDataGate } from "@repo/shared/components/EntityDataGate/entityDataGate";
+import { WorkflowsView } from "@repo/shared/views/WorkflowsView/workflowsView";
 import { type GetStaticProps } from "next";
 import { type JSX } from "react";
