+++ site-config/brc-analytics/local/index/workflow/types.ts
@@ -1,7 +1,7 @@
 import {
   type WorkflowEntity as BaseWorkflowEntity,
   type WorkflowAssembly,
-} from "@/views/WorkflowsView/types";
+} from "@repo/shared/views/WorkflowsView/types";
 
 export type WorkflowEntity = Omit<BaseWorkflowEntity, "assembly"> & {
   assembly: WorkflowAssembly;
